// libs/shared/common/src/lib/enums/colors.enum.ts
export enum AllColors {
  Red = 'Red',
  White = 'White',
  Blue = 'Blue',
  Black = 'Black',
  Green = 'Green',
  Multicolor = 'Multicolor',
  Colorless = 'Colorless',
  Land = 'Land',
  Azorius = 'Azorius',
  Dimir = 'Dimir',
  Rakdos = 'Rakdos',
  Gruul = 'Gruul',
  Selesnya = 'Selesnya',
  Orzhov = 'Orzhov',
  Golgari = 'Golgari',
  Simic = 'Simic',
  Izzet = 'Izzet',
  Boros = 'Boros',
  Bant = 'Bant',
  Esper = 'Esper',
  Grixis = 'Grixis',
  Jund = 'Jund',
  Naya = 'Naya',
  Mardu = 'Mardu',
  Temur = 'Temur',
  Abzan = 'Abzan',
  Jeskai = 'Jeskai',
  Sultai = 'Sultai',
}

/* 

 White + Blue = Azorius (guild on Ravnica), Ojutai (Dragonlord on Tarkir)
 Blue + Black = Dimir (guild on Ravnica), Silumgar (Dragonlord on Tarkir)
 Black + Red = Rakdos (guild on Ravnica), Kolaghan (Dragonlord on Tarkir)
 Red + Green = Gruul (guild on Ravnica), Atarka (Dragonlord on Tarkir)
 Green + White = Selesnya (guild on Ravnica), Dromoka (Dragonlord on Tarkir)


 White + Black = Orzhov (guild on Ravnica), Silverquill (College on Strixhaven)
 Black + Green = Golgari (guild on Ravnica), Witherbloom (College on Strixhaven)
 Green + Blue = Simic (guild on Ravnica), Quandrix (College on Strixhaven)
 Blue + Red = Izzet (Guild on Ravnica), Prismari (College on Strixhaven)
 Red + White = Boros (guild on Ravnica), Lorehold (College on Strixhaven)


 Green + White + Blue = Bant (shard of Alara), Brokers (New-Capenna's crime family)
 White + Blue + Black = Esper (shard of Alara), Obscura (New-Capenna's crime family)
 Blue + Black + Red = Grixis (shard of Alara), Maestros (New-Capenna's crime family)
 Red + Green + Black = Jund (shard of Alara), Riveteers (New-Capenna's crime family)
 Red + Green + White = Naya (shard of Alara), Cabaretti (New-Capenna's crime family)

 Red + white + black = Mardu (clan on Tarkir), Savai (Ikoria triome), Oros (dragon from Planar Chaos) or Dega (from Degavolver)
 Green + blue + red = Temur (clan on Tarkir), Ketria (Ikoria triome), Intet (dragon from Planar Chaos) or Ceta (from Cetavolver)
 White + black + green = Abzan (clan on Tarkir), Indatha (Ikoria triome), Teneb (dragon from Planar Chaos) Necra (from Necravolver), Junk citation, or Doran citation
 Blue + Red + White = Jeskai (clan on Tarkir), Raugrin (Ikoria triome), Numot (dragon from Planar Chaos) or Raka (from Rakavolver)
 Black + green + blue = Sultai (clan on Tarkir), Zagoth (Ikoria triome), Vorosh (dragon from Planar Chaos) or Ana (from Anavolver)

*/

export enum CommonColors {
  Red = 'Red',
  White = 'White',
  Blue = 'Blue',
  Black = 'Black',
  Green = 'Green',
}